+++ src/components/store/nvmestore/unit_test/test-throughput.cpp
@@ -95,8 +95,14 @@ TEST_F(KVStore_test, Instantiate)
   ASSERT_TRUE(comp);
   IKVStore_factory * fact = (IKVStore_factory *) comp->query_interface(IKVStore_factory::iid());
 
-  // this nvme-store use a block device and a block allocator
-  _kvstore = fact->create("owner","name", opt.pci);
+  std::map<std::string, std::string> params;
+  params["owner"] = "testowner";
+  params["name"] = "testname";
+  params["pci"] = opt.pci;
+  params["pm_path"] = "/mnt/pmem0/";
+  unsigned debug_level = 0;
+
+  _kvstore = fact->create(debug_level, params);
 #else
   Component::IBase * comp = Component::load_component("libcomanche-storefile.so",
                                                       Component::filestore_factory);